review: clean & merge files related to grid
