+++ src/components/Grid/index.ts
@@ -8,15 +8,15 @@
 /*export { AgGrid } from './AgGrid';
 export type { AgGridRef } from './AgGrid';*/
 
-export { GridButtonAdd } from './buttons/ButtonAdd';
-export type { GridButtonAddProps } from './buttons/ButtonAdd';
-//export { GridButtonDelete, GridButtonDeleteProps } from './buttons/ButtonDelete';
-//export { GridButtonRefresh, GridButtonRefreshProps } from './buttons/ButtonRefresh';
-
 export { default as NoRowsOverlay } from './NoRowsOverlay';
-export { GridColumnTypes } from './GridFormat';
-export { GridTable } from './GridTable';
-export type { GridTableProps } from './GridTable';
-
-export { default as DataGrid } from './DataGrid';
-export type { DataGridProps, DataGridRef } from './DataGrid';
+export {
+    GridTable,
+    GridColumnTypes,
+    GridButton,
+    GridButtonDelete,
+} from './GridTable';
+export type {
+    GridTableProps,
+    GridTableRef,
+    GridButtonProps,
+} from './GridTable';
