+++ src/translations/en.json
@@ -16,11 +16,9 @@
     "table.id": "ID",
     "table.error.retrieve": "Error while retrieving data",
     "table.toolbar.add": "Add",
-    "table.toolbar.add.label": "Add element",
-    "table.toolbar.add.tooltip": "Add an element",
+    "table.toolbar.add.label": "Add an element",
     "table.toolbar.delete": "Delete",
-    "table.toolbar.delete.label": "Delete selection",
-    "table.toolbar.delete.tooltip": "Delete selected element(s)",
+    "table.toolbar.delete.label": "Delete selected element(s)",
     "table.bool.yes": "Yes",
     "table.bool.no": "No",
     "table.bool.unknown": "Unknown",
@@ -32,8 +30,7 @@
     "users.table.error.delete": "Error while deleting user",
     "users.table.error.add": "Error while adding user",
     "users.table.toolbar.add": "Add user",
-    "users.table.toolbar.add.label": "Add user",
-    "users.table.toolbar.add.tooltip": "Add a user",
+    "users.table.toolbar.add.label": "Add a user",
     "users.form.title": "Add a user",
     "users.form.content": "Please fill in new user data.",
     "users.form.field.username.label": "User ID"
