+++ src/translations/fr.json
@@ -16,13 +16,11 @@
     "table.id": "ID",
     "table.error.retrieve": "Erreur pendant la récupération des données",
     "table.toolbar.add": "Ajouter",
-    "table.toolbar.add.label": "Ajouter élément",
-    "table.toolbar.add.tooltip": "Ajouter un élément",
+    "table.toolbar.add.label": "Ajouter un élément",
     "table.toolbar.delete": "Supprimer",
-    "table.toolbar.delete.label": "Supprimer sélection",
-    "table.toolbar.delete.tooltip": "Supprimer le(s) élément(s) sélectionné(s)",
+    "table.toolbar.delete.label": "Supprimer le(s) élément(s) sélectionné(s)",
     "table.bool.yes": "Oui",
-    "table.bool.no": "No",
+    "table.bool.no": "Non",
     "table.bool.unknown": "Inconnu",
 
     "users.title": "Liste des utilisateurs",
@@ -32,8 +30,7 @@
     "users.table.error.delete": "Erreur pendant la suppression de l'utilisateur",
     "users.table.error.add": "Erreur pendant l'ajout de l'utilisateur",
     "users.table.toolbar.add": "Ajouter utilisateur",
-    "users.table.toolbar.add.label": "Ajout utilisateur",
-    "users.table.toolbar.add.tooltip": "Ajouter un utilisateur",
+    "users.table.toolbar.add.label": "Ajouter un utilisateur",
     "users.form.title": "Ajouter un utilisateur",
     "users.form.content": "Veuillez renseigner les informations de l'utilisateur.",
     "users.form.field.username.label": "ID utilisateur"
